requirements.txt
python-decouple
Django==5.0.6
djangorestframework
django-cors-headers
openai

